ch04/KnowledgeAgent/requirements.txt
openai>=1.3.7
PyPDF2>=3.0.1
numpy>=1.24.3
pandas>=2.1.4
python-dotenv>=1.0.0
tiktoken>=0.5.2
requests>=2.31.0 